refactor base actions
